remove core.types.chain
