+++ core/builtins/message/chain.py
@@ -22,12 +22,14 @@ from core.builtins.message.internal import (
 )
 from core.builtins.utils import Secret
 from core.logger import Logger
-from core.types.message import MessageChain as MessageChainT
 
 from core.utils.http import url_pattern
 
 
-class MessageChain(MessageChainT):
+class MessageChain:
+    """
+    消息链。
+    """
     def __init__(
         self,
         elements: Union[
@@ -43,6 +45,9 @@ class MessageChain(MessageChainT):
             I18NContext,
         ] = None,
     ):
+        """
+        :param elements: 消息链元素
+        """
         self.value = []
         if isinstance(elements, ErrorMessage):
             elements = str(elements)
@@ -131,6 +136,9 @@ class MessageChain(MessageChainT):
 
     @property
     def is_safe(self):
+        """
+        检查消息链是否安全。
+        """
         def unsafeprompt(name, secret, text):
             return f'{name} contains unsafe text "{secret}": {text}'
 
@@ -182,6 +190,9 @@ class MessageChain(MessageChainT):
         return True
 
     def as_sendable(self, msg: 'MessageSession' = None, embed=True):
+        """
+        将消息链转换为可发送的格式。
+        """
         locale = None
         if msg:
             locale = msg.locale.locale
@@ -223,6 +234,9 @@ class MessageChain(MessageChainT):
         return value
 
     def to_list(self, locale="zh_cn", embed=True, msg: 'MessageSession' = None):
+        """
+        将消息链转换为列表。
+        """
         value = []
         for x in self.value:
             if isinstance(x, Embed) and not embed:
@@ -248,16 +262,34 @@ class MessageChain(MessageChainT):
             )
         return value
 
+    def from_list(self, lst: list):
+        """
+        从列表构造消息链。
+        """
+        raise NotImplementedError
+
     def append(self, element):
+        """
+        添加一个消息链元素到末尾。
+        """
         self.value.append(element)
 
     def remove(self, element):
+        """
+        删除一个消息链元素。
+        """
         self.value.remove(element)
 
     def insert(self, index, element):
+        """
+        在指定位置插入一个消息链元素。
+        """
         self.value.insert(index, element)
 
     def copy(self):
+        """
+        复制一个消息链。
+        """
         return MessageChain(self.value.copy())
 
     def __str__(self):
@@ -265,6 +297,28 @@ class MessageChain(MessageChainT):
 
     def __repr__(self):
         return self.__str__()
+
+    def __iter__(self):
+        return iter(self.value)
+
+    def __add__(self, other):
+        if isinstance(other, MessageChain):
+            return MessageChain(self.value + other.value)
+        else:
+            return MessageChain(self.value + other)
+
+    def __radd__(self, other):
+        if isinstance(other, MessageChain):
+            return MessageChain(other.value + self.value)
+        else:
+            return MessageChain(other + self.value)
+
+    def __iadd__(self, other):
+        if isinstance(other, MessageChain):
+            self.value += other.value
+        else:
+            self.value += other
+        return self
 
 
 def match_kecode(text: str) -> List[Union[Plain, Image, Voice, Embed]]:
